feeds: move feed processor id to config associate type
